requirements.txt
pandas==2.0.3
numpy==1.24.4 
scikit-learn==1.3.2
scipy==1.10.1
xgboost==2.1.1
Flask==2.2.3
gunicorn==23.0.0
pyYAML
ensure
requests
-e .